api_cli/requirements-pip3.txt
click
psycopg2-binary==2.9.6
prettytable
